pyproject.toml
[tool.poetry]
name = "moonlight"
version = "0.2.0"
description = "Wizard101 traffic decoder"
authors = ["Ethan Zeigler <ethanzeigler@gmail.com>", "Kronos Authors <>"]
license = "GPL-3.0-only"
readme = "README.md"
repository = "https://github.com/kronos-project/moonlight"
include = [
    "moonlight",
    "res",
    "test",
    "poetry.lock",
    "pyproject.toml"
]

[tool.poetry.dependencies]
python = ">= 3.9"
# https://github.com/secdev/scapy/issues/3246 fix isn't on pypi
scapy = {git = "https://github.com/secdev/scapy.git", rev = "master"}
printrospector = {git = "https://gitlab.com/vale_/printrospector", rev = "main"}
click = "^8.0.4"
colorama = "^0.4.4"
PyYAML = "^6.0"
yamlloader = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^22.3.0"
pre-commit = "^2.17.0"
ipython = "^8.1.1"
ipykernel = "^6.9.2"
snakeviz = "^2.1.1"

[tool.poetry.scripts]
moonlight = "moonlight.__main__:cli_cmd"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "line-too-long", # We'll let black handle this. There are cases when long lines are okay.
    "missing-docstring" # Once stable I'll be more worried about this
]
